Made run in Game run on variable change
run in Game ha been renamed to try_run, which is called by a new method
called update (in Game). Application now calls update on Game.
In the future, we might want to move out the update method to a higher
up class, say Model, that in turn calls the relevant methods on any
aspect of the model (such as Game, Lobby).