I Assets now load
Soved the bug where asssets would not load as well as creating a spritebatch would crash the game. The only place where data could be fetched is inside the create() method of a aplication, by doing so and settign a static feild to this the data can be shared with the rest of the view.