View now works
I had to fix the fact that there was publishers publesing in games constructor, this was problematic since then there was no time for the app to subscribe to the publishers, since they are created in the games constructor. I also added code for sending ICard and showing the topcard in teh view.