Bug fix
I have now fixed a bug where the cards where not regester with an id when an equal card allready is regesterd. However it was still sent to teh client resulting in clinets getting invalid cards to play with. The bug orgininated in the use of BiHashMap in ModelApplication since it was assumed two equal cards could both be keys in a hashmap. Thus, the fix was to make it so.